+++ types/types.go
@@ -65,6 +65,7 @@ type Config struct {
 	ETCDCPU                     float64
 	ETCDMEM                     float64
 	M3SStatus                   M3SStatus
+	MesosSandboxVar             string
 }
 
 // M3SStatus store the current TaskState of the M3s services